Fixed race condition on Win poll
It was possible that a thread would read resync as false before another thread set it to true and subsequently register a waker after the other thread set resync to true and already tried calling AtomicWaker::wake().

To fix this issue, we first register the waker, and only then we check the resync value. If resync is true and we didn't have to register the waker, we just take it back.